Special-case parsing of boolean values.
Fix issue 5 (continue_scan can't be set to False) by special-casing
boolean values in turnstile.tools.parse_limit_node().  If the value
is specified as a sequence of digits, the string value is coerced to
int, then coerced to bool; otherwise, the strings "t", "true", "on",
"y", and "yes" are converted to True and the strings "f", "false",
"off", "n", and "no" are converted to False.  Any other string
yields a warning.