Refactor limit object construction
Refactors limit object construction to decouple the control daemon
from the middleware.  Introduces the LimitData class, which acts as
a go-between for passing raw limit data taken from the database to
the middleware, which is then responsible for constructing the
routes for it.  This will allow a later refactoring which will
permit the ControlDaemon to run in a separate process; in this
case, the LimitData subclass used by the middleware will know how
to get the limit data from the LimitData subclass used by the
ControlDaemon.