Use function registration for commands
Pull command implementations out of the ControlDaemon class and
switch to a function registration methodology.  This removes the
ugly hack that all non-commands in ControlDaemon must begin with
an underscore, and will make the ControlDaemon easier to
understand, and to extend with support for multiprocess.  This
also makes it easy for third parties to define new ControlDaemon
commands, if desired.