Remove unneeded underscores
After refactoring command functions out of the ControlDaemon,
the '_' prefix is no longer necessary.  Remove the '_' prefix
from ControlDaemon everywhere it was used on something meant to
be public (the _commands map and _register() continue to use
the prefix, since they really are meant to be internal).  Also
corrects the middleware to get the LimitData using get_limits()
instead of using a direct reference; this will allow the multi-
process version of ControlDaemon to return a different object.