Refactor configuration
Refactors configuration by creating a new Config class.  Also
augments configuration: a "config" parameter can be provided in
the paste.ini to specify an alternate configuration file.
Unifies configuration needed by the Turnstile middleware and
the tools, with backwards-compatibility.

Note that the new configuration object is available in the WSGI
environment, in the 'turnstile.conf' value; this should be
contrasted with the 'turnstile.config' value, which is now
deprecated.