Add BucketLoader
Adds the BucketLoader, which is a specialized class to reconstruct a
bucket from a list of specialized records which document how that
bucket has been altered.  This algorithm is implemented in an
__init__() because several different outputs are needed to implement
the full scalability change.  This commit also includes partial unit
tests on BucketLoader, using mock.  (Future work will convert the
entire test suite over to using mock in preference to mox.)