Begin the process of reorganizing and rewriting the test suite
Pull the tests for all the new limits.py classes into a 'unit'
subdirectory.  These tests all use mock instead of the older
(and harder to use) mox.  Eventually, the full test suite will
be reimplemented in terms of mock, and unit tests will be
placed in the 'unit' subdirectory of the 'tests' directory,
allowing for the eventual addition of functional tests.