Add entrypoint search for middleware
Currently, middleware.turnstile_filter() looks up an alternative
TurnstileMiddleware class using utils.import_class(), only if a
colon exists in the name.  This change adds a search for a given
name using the utils.find_entrypoint() with an entrypoint group
of "turnstile.middleware".  Future plans entail adding an
alternative way of changing the format_delay() output, so the
'turnstile' configuration option may be deprecated in the
future.

This change also updates and moves the TestTurnstileFilter class
to the new test_middleware.py as part of the test suite update/
rewrite.