Refactor import_class() and find_entrypoint()
There was a lot of duplicated code all over the place because of the
way import_class() and find_entrypoint() were being used.  This
refactors all of that code into find_entrypoint(), reducing the
duplication.  Note that this breaks test_database.py, because that
file is in the process of being rewritten.