Eliminate TurnstileRedis
Eliminate the TurnstileRedis class by promoting the added methods
(limit_update() and command()) to functions.  This also allows the
Redis client class to be set in configuration, which might be
necessary to use some forms of Redis sharding.  Note that
test_database.py is not updated; it is due for a rewrite (see next
commit).