Add postprocessor
Some extension tools may want to use the ScriptAdaptor and
related decorators from tools.py, and may have need to post-
process the result from the underlying function, e.g., to
display results.  This adds a postprocessor concept to
accommodate this.  Also, the docstring on
ScriptAdaptor._add_preprocessor() was fixed to no longer
refer to the 'insert' argument or mention a return value.