Ignore empty configuration values for extra database arguments
Turnstile's functionality to allow the control daemon and the
compactor daemon to use different Redis configuration provides
one problem: values that are set in the "[redis]" section are
inherited by all sections, even when that is not desired.  To
combat this, we now allow an empty value to completely delete
the key from the redis configuration in Config.get_database().